Fix all clippy warnings, add integration tests, health endpoints, PTP integration
Clippy fixes (14 warnings -> 0):
- Doc comment formatting in servo.rs (module-level //! comments)
- Replace manual % checks with .is_multiple_of()
- Replace manual range checks with .contains()
- Replace .max().min() with .clamp()
- Remove unnecessary i64 casts in phc.rs
- Collapse nested if statements
- Add Default impl for MetricsExporter

Integration tests (32 new tests in tests/):
- ntp_roundtrip.rs: End-to-end NTP client/server over loopback UDP
- marzullo_integration.rs: Full filter->select pipeline with 5 sources
- nts_handshake.rs: Complete NTS-KE flow with cookies and AEAD
- config_loading.rs: TOML parsing with defaults and validation
- ptp_sync.rs: PTP message exchange, BMCA, E2E delay computation

Health/readiness endpoints:
- /health returns 200 (liveness probe)
- /ready returns 200 after first successful sync, 503 before
- Arc<AtomicBool> readiness flag set by selection loop

PTP daemon integration:
- PTP node task: joins multicast, processes Sync/FollowUp/DelayReq/DelayResp
- PTP SourceMeasurements feed into same Marzullo pipeline as NTP
- Hybrid NTP+PTP source selection

+++ crates/rtime-clock/src/phc.rs
@@ -136,7 +136,7 @@ impl Clock for PhcClock {
         }
 
         let nanos = offset.to_nanos();
-        let total_ns = ts.tv_sec as i64 * 1_000_000_000 + ts.tv_nsec as i64 + nanos;
+        let total_ns = ts.tv_sec * 1_000_000_000 + ts.tv_nsec + nanos;
 
         if total_ns < 0 {
             return Err(ClockError::Os(std::io::Error::new(
